Add files via upload
By analyzing audio features of ‘Viral 50’ Charts from 2019 to 2020, analyze the change in the features of trending music in the countries where the COVID-19 hit the hardest and the least. (assuming the audio features we select to analyze reflects people’s musical demand and sentiment /=reason for choosing to listen to the song) (https://spotifycharts.com/viral/ad/weekly/latest). Furthermore, implement sentiment analysis using the lyrics of the tracks in the dataset.

After cleaning up the initial data, we retrieved audio features data using the Spotify API. After adding audio features to the initial dataset, proceeded with the analysis.